Add files via upload
The folder contains the dataset for multiple linear regression and the code to implement it. There's a portion in it i.e. 'Building Optimal Model' which is the part that can be used to eliminate the entities that have too much effect on the profit or simply which are not optimal.